+++ src/core/arithmetic/modular/residues/index.ts
@@ -1,0 +1,3 @@
+export * from './isCoprime';
+export * from './findResidue';
+export * from './totient';